make GroupSubmissionsWrapper reuseable, add basic GroupSubmissionsPage
