+++ client/src/constants/queries.js
@@ -115,7 +115,9 @@ export const GET_CHALLENGE_GROUP = gql`
 export const GET_GROUP_SUBMISSIONS_COUNT = gql`
   query challengeGroup($challengeGroupId: ID!) {
     challengeGroup(challengeGroupId: $challengeGroupId) {
+      id
       challenges {
+        id
         submissions {
           id
         }
