+++ client/src/constants/routes.js
@@ -3,6 +3,7 @@ export const DASHBOARD = '/dashboard';
 export const LOGIN = '/login';
 export const SIGNUP = '/signup';
 export const CHALLENGE_GROUPS = '/groups';
+export const CHALLENGE_GROUP_SUBMISSIONS = '/groups/:id/submissions';
 export const CHALLENGE = '/challenge/:id';
 export const CHALLENGE_SETTINGS = '/challenge/:id/settings';
 export const ADD_SUBMISSION = '/challenge/:id/new';
